BustAMovePanel::Poll: mBeatCount >= 3 is a signed compare (99.5 -> 99.6)
Restores the semantically-correct form. The source carried an
`(unsigned int)mBeatCount >= 3` cast to force `cmplwi`; the target uses a
signed `cmpwi cr6, r11, 0x3`, so the cast was the artifact, not the fix.

This edit had been measured twice before and rejected both times:
  - on the 98.6% baseline: 98.6 -> 98.3 (-0.3). It fixed idx 129 but the
    kBAMState_ShowMoveSequence preheader at idx 206-210 degraded from two
    'replace' rows to four insert/delete rows.
  - on the 98.8% state (after the isPlayer0Pink inversion): 98.8 -> 98.5,
    same -0.3 for the same reason.

Both regressions were the preheader ordering bug fixed in the previous commit
(0a7dfc2b, 'scores' local removed). With that cluster matching, removing the
cast is free and pays its own +0.1: idx 129 becomes the only change and the
preheader stays aligned. Residual 6 -> 5.

Lesson worth keeping: a correct edit can measure as a regression while an
unrelated defect is still masking it. Park such edits and retry after every
subsequent landed fix rather than discarding them on the first measurement.

Behaviourally the two forms were already identical here — mBeatCount is reset
to 0 and only ever incremented, so it is never negative.

+++ src/lazer/game/BustAMovePanel.cpp
@@ -1535,7 +1535,7 @@ void BustAMovePanel::Poll() {
         mRecordSkelIdx = skelIdx;
     }
 
-    if (mState == kBAMState_Recording && (unsigned int)mBeatCount >= 3) {
+    if (mState == kBAMState_Recording && mBeatCount >= 3) {
         mDancerTakeScore = mRecorder->GetScore(skelIdx, 0, mRecordScore, true);
         mCurrentMoveScore = mRecorder->GetScore(skelIdx, 1, mRecordScore, false);
         mRecordScore += TheTaskMgr.DeltaUISeconds();
